Move TaskStatus into ui package
